feat: add first Up/Down detection
Using a simple but yet effective means:
If a packet is sent, save it's unix time and when a packet
is received, save it's unix time too. Once per second,
we'll check if the time between rx/tx is greater than threshold.
If it is, we'll mark it as an outage. This can be configured
and is still WIP.

For example, we would be able to not take into account packet
that don't ask for a reply (think UDP or Icmp for some).

Signed-off-by: Martichou <m@rtin.fyi>